phone-backend/requirements.txt
asterisk-ami
bottle
redis
